Remove sample support and content component and remove css accordingly
